preparing features for nodes
This version creates multiple test cases where we prepare test data by
expanding the graph by a fixed number of layers. Each node is recursed
N number of times to look at other nodes.

Next steps could be to count the number of incoming fields and types,
also to split out tables of different types/morphs/forms.

We want to try and train a netork to look at each field and see what
other field is significant to it, where there is a correlation.